Assignments 1 and 2
Assignments 1 and 2 for COMP219, including the documentation of my projects. Assignment 1, to create a neural network using KNN and train it on the digits dataset (simplified MNIST dataset), after training the scikit-learn library function on the digits dataset; compare the two models. Assignment 2, to create a neural network and train it on the digits dataset (with and without a convolutional layer). Then, compare the accuracy of the two models, along with the implementation on your own (KNN model from A1).